components/continue_reading_card: Improve card layout and presentation
Use skeletons when loading and drop loading the most recent history when it gets built. This is used on surfaces that infinitely scroll and as such gets initialized very frequently. This'll make this card more elegant.

Since we now load most recent history on app start, drop it here as well. This doesn't affect loading when it changes since the card consumes ReadHistoryProvider anyway.

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>