recommendations: Refactor recommendations to be extensible
Make the recommendation engines an abstract class so potential recommenders can define common properties for all recommendations, as well as provide their own type of recommendations. This will allow more types of recommendations (such as series and collections) to be added in the future.

For now, let's re-implement all existing recommendation engines with the new mechanism.

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>