expressive: Introduce a set of Material 3 Expressive widgets
These are reimplemented versions of the Material Scaffold, AppBar, and NestedScrollViews meant for use with Archiverse which uses a Material 3 Expressive-adjacent design. Flutter upstream is currently reorganizing how UIs are handled in Flutter and as such M3E (and Liquid Glass) are being put on hold at the moment. However, we need this a lot right now so let's implement it ourselves for now.

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>