work/series: Improve loading of items
We hit a race because of the request throttling and it results to the page fetching the author image before the item itself, making wait times longer. Let's fix that by deferring loading the user image while the page itself is still loading.

While we're at it, commonize the component.

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>