database: Make read histories more intuitive
Instead of keeping one history for a work (because a user can read it on multiple days and times and this completely breaks that expectation since it'll reorganize histories), let's allow saving multiple histories for a work given that they are different sessions.

We'll have to update the UI to actually take advantage of this but that is the right thing to do anyway.

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>