reader: Refactor the reader view
Completely clean up this messy piece of code. This makes each component modular so they're easier to maintain (and even add to more activities if we do ever need to).

This also (finally) adds a proper way to save and restore scrolling position, allowing the app to go back exactly where the user left off for a given read history.

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>